requirements.txt
altair==5.5.0
altex==0.2.0
appnope==0.1.4
asn1crypto==1.5.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
boto3==1.40.34
botocore==1.40.34
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cloudpickle==3.1.1
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.0
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
entrypoints==0.4
exceptiongroup==1.3.0
executing==2.2.0
Faker==37.8.0
favicon==0.7.0
filelock==3.19.1
fonttools==4.60.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.1.10
htbuilder==0.9.0
huggingface-hub==0.35.0
idna==3.10
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lxml==6.0.1
Markdown==3.9
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==2.5.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
polars==1.33.1
prometheus_client==0.23.1
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pyOpenSSL==25.3.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
s3transfer==0.14.0
safetensors==0.6.2
scipy==1.15.3
segmentation_models_pytorch==0.5.0
six==1.17.0
smmap==5.0.2
snowflake-connector-python==3.17.3
snowflake-snowpark-python==1.39.0
sortedcontainers==2.4.0
soupsieve==2.8
st-annotated-text==4.0.2
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.49.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.7.8
streamlit-image-coordinates==0.4.0
streamlit-keyup==0.3.0
streamlit-notify==0.3.1
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamlit_faker==0.0.4
sympy==1.14.0
tenacity==9.1.2
timm==1.0.19
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.201
ultralytics-thop==2.0.17
urllib3==2.5.0
uv==0.8.18
validators==0.35.0
wcwidth==0.2.13
